Update README.md
This repository contains a solution for the Zindi Animal Classification Challenge, which involves a dataset consisting of 13,999 animal images in the training set and 5,000 images in the test set. The dataset comprises two types of animals: zebras and elephants.

The data for this challenge was sourced from the Data Repository for the University of Minnesota under a creative commons license. It originates from a study titled "Identifying Animal Species in Camera Trap Images using Deep Learning and Citizen Science" and includes images downloaded from Zooniverse, all resized to 330x330 pixels.

The camera traps were strategically placed in Tanzania within the Serengeti National Park, triggered by a combination of infrared and motion sensors. They captured three images per trigger event, with a one-minute block between triggers.

The main objective of this challenge is to estimate the probability that an image contains a zebra. Participants are tasked with providing likelihood estimates for zebras in each unique image, with values ranging between 0 and 1. Additional information about the dataset can be found at https://www.snapshotserengeti.org/ and https://conservancy.umn.edu/handle/11299/199819.